refactor setup project
