--- src/main/kotlin/logic/entities/State.kt
@@ -1,7 +1,0 @@
-package logic.entities
-
-enum class State(val title: String) {
-    TODO("Todo"),
-    IN_PROGRESS("In Progress"),
-    DONE("Done")
-}
