+++ src/main/kotlin/logic/models/EntityType.kt
@@ -1,4 +1,4 @@
-package logic.entities
+package logic.models
 
 enum class EntityType {
     PROJECT, TASK
