+++ src/main/kotlin/logic/models/Project.kt
@@ -5,5 +5,4 @@ import java.util.*
 data class Project(
     val id: UUID,
     val name: String,
-    val tasks: List<Task>
 )
