+++ src/main/kotlin/logic/models/Role.kt
@@ -1,4 +1,4 @@
-package logic.entities
+package logic.models
 
 enum class Role {
     ADMIN, MATE
