+++ src/main/kotlin/logic/models/State.kt
@@ -1,0 +1,9 @@
+package logic.entities
+
+import java.util.UUID
+
+data class State(
+    val id: UUID,
+    val title: String,
+    val projectId: UUID,
+)
