+++ src/main/kotlin/logic/models/Task.kt
@@ -6,5 +6,6 @@ data class Task(
     val id: UUID,
     val title: String,
     val description: String,
-    val state: State
+    val projectId: UUID,
+    val stateId: UUID,
 )
