+++ src/main/kotlin/presentation/PlanMateConsoleUI.kt
@@ -3,7 +3,7 @@ package presentation
 import presentation.io.ConsoleIO
 
 class PlanMateConsoleUI(
-    private val consoleIO: ConsoleIO
-) {
+    consoleIO: ConsoleIO
+) : ConsoleIO by consoleIO {
 
 }
