+++ src/main/kotlin/utilities/HashingExtension.kt
@@ -1,10 +1,1 @@
 package utilities
-
-import java.security.MessageDigest
-
-@OptIn(ExperimentalStdlibApi::class)
-fun String.md5(): String {
-    val md = MessageDigest.getInstance("MD5")
-    val digest = md.digest(this.toByteArray())
-    return digest.toHexString()
-}
